Azure Lab 04 Implement Virtual Networking
Summary and Resources
Azure Virtual Network is the fundamental building block for your private network in Azure. VNet enables many types of Azure resources, such as Azure Virtual Machines, to securely communicate with each other, the internet, and on-premises networks. VNet is similar to a traditional network that you'd operate in your own data center, but brings with it additional benefits of Azure's infrastructure such as scale, availability, and isolation.

Azure IP addressing is critical to ensuring resources are accessible. Private IP addresses to communicate between resources in Azure. Public IP addresses enable Azure resources to be accessible directly from the internet..

You should now be able to:

Identify features and usage cases for subnets and subnetting.

Create a virtual network with subnetting.

Identify usage cases for private and public IP addresses.

Create and determine which resources require public and private IP addresses.